+++ nodes/ble/00-bledevice.js
@@ -1,0 +1,8 @@
+module.exports = function(RED) {
+    'use strcit';
+    function bleDevice(n){
+        RED.nodes.createNode(this, n);
+        this.address = n.address;
+    }
+    RED.nodes.registerType("bledevice", bleDevice);
+};
