+++ nodes/ble/05-read.html
@@ -1,0 +1,56 @@
+<script type="text/x-red" data-template-name="ble read">
+    <div class="form-row">
+        <label for="node-input-device"><i class="fa fa-info"></i> <span data-i18n="ble-read.label.device"></span></label>
+        <input type="text" id="node-input-device">
+    </div>
+
+    <div class="form-row">
+        <label for="node-config-input-service"><i class="fa fa-info"></i> <span data-i18n="ble-read.label.service"></span></label>
+        <input type="text" id="node-input-service">
+    </div>
+
+    <div class="form-row">
+        <label for="node-config-input-characteristic"><i class="fa fa-info"></i> <span data-i18n="ble-read.label.characteristic"></span></label>
+        <input type="text" id="node-input-characteristic">
+    </div>
+
+    <div class="form-row">
+        <label for="node-input-name"><i class="fa fa-tag"></i> <span data-i18n="node-red:common.label.name"></span></label>
+        <input type="text" id="node-input-name" data-i18n="[placeholder]node-red:common.label.name">
+    </div>
+</script>
+
+<script type="text/javascript">
+    RED.nodes.registerType('ble read', {
+        category: 'ble',
+        color: '#4d94ff',
+        defaults: {
+            name: {
+                value: ""
+            },
+            device: {
+                value: "",
+                type: "bledevice"
+            },
+            service: {
+                value:"",
+                required: true
+            },
+            characteristic: {
+                value: "",
+                required: true
+            }
+        },
+        inputs: 1,
+        outputs: 1,
+        align: 'right',
+        icon: "ble.png",
+        label: function () {
+            return this.name || this._("ble-read.ble-read");
+        },
+        oneditprepare: function() {
+        },
+        oneditsave: function() {
+        }
+    });
+</script>
