+++ nodes/ble/locales/en-US/00-bledevice.html
@@ -1,0 +1,3 @@
+<script type="text/x-red" data-help-name="bledevice">
+    <p>You can specify the address of ble device to change it.</p>
+</script>
