+++ nodes/ble/locales/en-US/00-bledevice.json
@@ -1,0 +1,8 @@
+{
+    "bledevice":{
+        "bledevice": "device",
+        "label":{
+            "address": "Address"
+        }
+    }
+}
