+++ nodes/ble/locales/en-US/01-notification.html
@@ -1,0 +1,16 @@
+<script type="text/x-red" data-help-name="ble notification">
+    <p>Outputs notifications from subscribed devices.</p>
+    <h3>Outputs</h3>
+    <dl class="message-properties">
+        <dt>payload.name <span class="property-type">string</span></dt>
+        <dd>Device name</dd>
+        <dt>payload.address <span class="property-type">string</span></dt>
+        <dd>Device address</dd>
+        <dt>payload.service <span class="property-type">string</span></dt>
+        <dd>Service UUID</dd>
+        <dt>payload.characteristic <span class="property-type">string</span></dt>
+        <dd>Characteristic UUID</dd>
+        <dt>payload.value <span class="property-type">string</span></dt>
+        <dd>Base64 encoded string of bytes</dd>
+    </dl>
+</script>
