+++ nodes/ble/locales/en-US/01-notification.json
@@ -1,0 +1,5 @@
+{
+    "ble-notification":{
+        "ble-notification": "notification"
+    }
+}
