+++ nodes/ble/locales/en-US/03-connect.json
@@ -1,0 +1,8 @@
+{
+    "ble-connect":{
+        "ble-connect": "connect",
+        "label":{
+            "device": "Device"
+        }
+    }
+}
