+++ nodes/ble/locales/en-US/04-disconnect.html
@@ -1,0 +1,12 @@
+<script type="text/x-red" data-help-name="ble disconnect">
+    <p>Disconnect from a Bluetooth LE device.</p>
+    <h3>Outputs</h3>
+    <dl class="message-properties">
+        <dt>payload.name <span class="property-type">string</span></dt>
+        <dd>Device name</dd>
+        <dt>payload.address <span class="property-type">string</span></dt>
+        <dd>Device address</dd>
+        <dt>payload.status <span class="property-type">string</span></dt>
+        <dd>status</dd>
+    </dl>
+</script>
