+++ nodes/ble/locales/en-US/04-disconnect.json
@@ -1,0 +1,8 @@
+{
+    "ble-disconnect":{
+        "ble-disconnect": "disconnect",
+        "label":{
+            "device": "Device"
+        }
+    }
+}
