+++ nodes/ble/locales/en-US/05-read.json
@@ -1,0 +1,10 @@
+{
+    "ble-read":{
+        "ble-read": "read",
+        "label":{
+            "device": "Device",
+            "service": "Service",
+            "characteristic": "Characteristic"
+        }
+    }
+}
