+++ nodes/ble/locales/en-US/06-write.json
@@ -1,0 +1,10 @@
+{
+    "ble-write":{
+        "ble-write": "write",
+        "label":{
+            "device": "Device",
+            "service": "Service",
+            "characteristic": "Characteristic"
+        }
+    }
+}
