+++ nodes/ble/locales/en-US/07-subscribe.json
@@ -1,0 +1,10 @@
+{
+    "ble-subscribe":{
+        "ble-subscribe": "subscribe",
+        "label":{
+            "device": "Device",
+            "service": "Service",
+            "characteristic": "Characteristic"
+        }
+    }
+}
