+++ nodes/ble/locales/en-US/08-unsubscribe.html
@@ -1,0 +1,16 @@
+<script type="text/x-red" data-help-name="ble unsubscribe">
+    <p>Unsubscribe to a particular service's characteristic.</p>
+    <h3>Outputs</h3>
+    <dl class="message-properties">
+        <dt>payload.name <span class="property-type">string</span></dt>
+        <dd>Device name</dd>
+        <dt>payload.address <span class="property-type">string</span></dt>
+        <dd>Device address</dd>
+        <dt>payload.service <span class="property-type">string</span></dt>
+        <dd>Service UUID</dd>
+        <dt>payload.characteristic <span class="property-type">string</span></dt>
+        <dd>Characteristic UUID</dd>
+        <dt>payload.status <span class="property-type">string</span></dt>
+        <dd>status</dd>
+    </dl>
+</script>
