+++ nodes/ble/locales/en-US/08-unsubscribe.json
@@ -1,0 +1,10 @@
+{
+    "ble-unsubscribe":{
+        "ble-unsubscribe": "unsubscribe",
+        "label":{
+            "device": "Device",
+            "service": "Service",
+            "characteristic": "Characteristic"
+        }
+    }
+}
