+++ package.json
@@ -52,7 +52,15 @@
       "volume": "./nodes/volume/volume.js",
       "intent": "./nodes/intent/intent.js",
       "sqlite": "./nodes/sqlite/sqlite.js",
-      "ble": "./nodes/ble/ble.js"
+      "ble-bledevice": "./nodes/ble/00-bledevice.js",
+      "ble-notification": "./nodes/ble/01-notification.js",
+      "ble-scan": "./nodes/ble/02-scan.js",
+      "ble-connect": "./nodes/ble/03-connect.js",
+      "ble-disconnect": "./nodes/ble/04-disconnect.js",
+      "ble-read": "./nodes/ble/05-read.js",
+      "ble-write": "./nodes/ble/06-write.js",
+      "ble-subscribe": "./nodes/ble/07-subscribe.js",
+      "ble-unsubscribe": "./nodes/ble/08-unsubscribe.js"
     }
   }
 }
